Version final con demo incluida

+++ AppCutAudio/AppCutAudio/Helpers/AudioDetector.cs
@@ -27,7 +27,7 @@ namespace AppCutAudio.Helpers
                 int read;
                 bool isRecording = false;
                 MemoryStream outputStream = new MemoryStream();
-                WaveFileWriter writer = null;
+                LameMP3FileWriter writer = null;
                 int trackNumber = 1;
 
                 while ((read = await Task.Run(() => reader.Read(buffer, 0, buffer.Length))) > 0)
@@ -39,42 +39,41 @@ namespace AppCutAudio.Helpers
                         if (abs > maxVolume) maxVolume = abs;
                     }
 
-                    // Comienza a grabar si el volumen supera el umbral y aún no está grabando
                     if (maxVolume > 0.1f && !isRecording)
                     {
                         isRecording = true;
-                        writer = new WaveFileWriter(new IgnoreDisposeStream(outputStream), reader.WaveFormat);
+                        writer = new LameMP3FileWriter(new IgnoreDisposeStream(outputStream), reader.WaveFormat, LAMEPreset.STANDARD);
                     }
 
-                    // Si está grabando y el volumen cae por debajo del umbral, detiene la grabación y guarda la pista
                     if (isRecording && maxVolume <= 0.1f)
                     {
                         isRecording = false;
                         writer?.Dispose();
                         writer = null;
 
-                        // Guarda el contenido de outputStream en un archivo
-                        await Task.Run(() => fileIO.WriteAllBytesAsync($"{pathSave}/pista_{trackNumber}.wav", outputStream.ToArray()));
-                        outputStream.SetLength(0); // Limpia el MemoryStream para la siguiente pista
-                        trackNumber++; // Incrementa el número de pista
+                        await Task.Run(() => fileIO.WriteAllBytesAsync($"{pathSave}/pista_{trackNumber}.mp3", outputStream.ToArray()));
+                        outputStream.SetLength(0);
+                        trackNumber++;
                     }
 
-                    // Escribe en el archivo si está grabando
                     if (isRecording)
                     {
-                        writer?.WriteSamples(buffer, 0, read);
+                        byte[] byteArray = new byte[read * sizeof(float)];
+                        Buffer.BlockCopy(buffer, 0, byteArray, 0, byteArray.Length);
+                        writer?.Write(byteArray, 0, byteArray.Length);
                     }
                 }
 
                 writer?.Dispose();
 
-                // Si quedó alguna pista sin guardar al final del archivo, guárdala ahora
                 if (outputStream.Length > 0)
                 {
-                    await Task.Run(() => fileIO.WriteAllBytesAsync($"{pathSave}/pista_{trackNumber}.wav", outputStream.ToArray()));
+                    await Task.Run(() => fileIO.WriteAllBytesAsync($"{pathSave}/pista_{trackNumber}.mp3", outputStream.ToArray()));
                 }
             }
         }
+
+
         private async Task ConvertWavToMp3(string wavFileName, string mp3FileName, ID3TagData iD3)
         {
             using (var reader = new WaveFileReader(wavFileName))
@@ -91,13 +90,23 @@ namespace AppCutAudio.Helpers
 
             foreach (string nombreArchivo in segmentos)
             {
-                string mp3File = nombreArchivo.Replace(".wav", ".mp3");
-                await ConvertWavToMp3(nombreArchivo, mp3File, iD3);
-                File.Delete(nombreArchivo);
+                string extencion = nombreArchivo.Split('.').ToList().LastOrDefault();
+                if (extencion == "wav")
+                {
+                    string mp3File = nombreArchivo.Replace(".wav", ".mp3");
+                    await ConvertWavToMp3(nombreArchivo, mp3File, iD3);
+                    File.Delete(nombreArchivo);
 
-                AudioFile item = new AudioFile(Path.GetFileName(mp3File), mp3File, iD3);
+                    AudioFile item = new AudioFile(Path.GetFileName(mp3File), mp3File, iD3);
 
-                segmentosMp3.Add(item);
+                    segmentosMp3.Add(item);
+                }
+                else if (extencion == "mp3")
+                {
+                    AudioFile item = new AudioFile(Path.GetFileName(nombreArchivo), nombreArchivo, iD3);
+                    if (item.NombreArchivo != "demo.mp3")
+                        segmentosMp3.Add(item);
+                }
             }
 
             return segmentosMp3;
